move Node and Iterator classes inside CLImpl, rename test methods, add gitignore on module level, make decomposition of search method
