+++ collection/pom.xml
@@ -11,5 +11,4 @@
 
     <artifactId>collection</artifactId>
 
-
 </project>
